backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-dotenv

# Data Validation
pydantic

# File Upload
python-multipart

# Data Processing (POS Excel/CSV)
pandas
openpyxl
Pillow

# Google Cloud Services
google-cloud-documentai
google-cloud-aiplatform
google-cloud-firestore
google-cloud-bigquery
google-cloud-storage
langchain
langchain-community
langchain-google-vertexai
chromadb

# Firebase Authentication
firebase-admin
